small changes to setProbabilityAndNormalise

+++ AHA/src/main/java/Learner/EmissionMatrix.java
@@ -137,7 +137,7 @@ public class EmissionMatrix extends CommonMatrix
     return Pair.with(mostProbableEmissionState, maxProbability);
   }
 
-  public void setProbabilityAndNormalise(int newProbability, HiddenState hiddenState, EmissionState emissionState){
+  public void setProbabilityAndNormalise(double newProbability, HiddenState hiddenState, EmissionState emissionState){
     int hiddenStateIndex = mapWarden.hiddenStateToHiddenStateIndex(hiddenState);
     int emissionStateIndex = mapWarden.emissionStateToEmissionStateIndex(emissionState);
 
@@ -150,7 +150,7 @@ public class EmissionMatrix extends CommonMatrix
 
     // normalise the rest of the values
     int valuesToNormalise = (matrix.getColumnDimension() * matrix.getRowDimension() - 1);
-    double valueToOffsetRest = diff / valuesToNormalise;
+    double valueToOffsetRest = diff / (double) valuesToNormalise;
     for (int row = 0; row < matrix.getRowDimension(); row++){
       for (int col = 0; col < matrix.getColumnDimension(); col++){
         if (row != emissionStateIndex && col != hiddenStateIndex) {
